+++ lib/rus3.rb
@@ -49,6 +49,7 @@ module Rus3
   require_relative "rus3/procedure/control"
   require_relative "rus3/procedure/write"
 
+  require_relative "rus3/ast"
   require_relative "rus3/parser"
   require_relative "rus3/evaluator"
   require_relative "rus3/printer"